Updates to a lot of stuff.  Main part: upgraded workbenches.
Changes up the inheritances so they all inherit off a base blueprint.  Added the v2 and v3 (A and B) versions of the workbenches.
Added auto-processing (cooking, tanning, smelting, etc.) recipes.
Added new feats for said v2 and v3 benches.
Added a limit to bench inventory (leaving them running for days shouldn't be a concern anymore - don't know if it ever was, but once you have 100 items in the inventory, it will cancel all remaining recipes and refuse to queue anymore).